Add toolbar menu to all screens
