+++ ViewModels/MainMenuViewModel.cs
@@ -10,39 +10,11 @@ namespace Jeffistance.ViewModels
     {
         MainWindowViewModel parent;
         public string Greeting => "0.01";
-        public string TheTruth => "Jeff ungae";
-
-        public ReactiveCommand<Control, Unit> AboutMenu { get; } 
+        public string TheTruth => "Jeff ungae"; 
 
         public MainMenuViewModel(MainWindowViewModel parent)
         {
             this.parent = parent;
-
-            AboutMenu = ReactiveCommand.Create<Control>(
-                (control) => {
-                    Window popupWindow = CreateAboutWindow();
-                    popupWindow.ShowDialog((Window)control.GetVisualRoot());
-                }
-            );
-        }
-
-        private Window CreateAboutWindow()
-        {
-            Window window = new Window()
-            {
-                Title = "About",
-                ShowInTaskbar = false,
-                Height = 200,
-                Width = 200,
-                WindowStartupLocation = WindowStartupLocation.CenterOwner
-            };
-            window.Content = new TextBlock {
-                Text = "Jeffistance is a very epic app made by epic people for epic people.",
-                TextWrapping = TextWrapping.Wrap,
-                FontSize = 24
-            };
-            
-            return window;
         }
 
         private void OnHostButtonClick()
