+++ Views/JeffWindow.xaml
@@ -3,11 +3,20 @@
         xmlns:vm="clr-namespace:Jeffistance.ViewModels;assembly=Jeffistance"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:views="clr-namespace:Jeffistance.Views"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
         x:Class="Jeffistance.Views.MainWindow"
         Icon="/Assets/avalonia-logo.ico"
-        Title="Jeffistance"
-        Content="{Binding Content}">
+        Title="Jeffistance">
+    
+    <DockPanel>
+        <views:ToolbarMenuView DockPanel.Dock="Top">
+            <views:ToolbarMenuView.DataContext>
+                <vm:ToolbarMenuViewModel/>
+            </views:ToolbarMenuView.DataContext>
+        </views:ToolbarMenuView>
+        <UserControl Content="{Binding Content}"/>
+    </DockPanel>
 
     <Design.DataContext>
         <vm:MainWindowViewModel/>
