+++ Views/MainMenuView.xaml
@@ -9,40 +9,6 @@
   <DockPanel
     Background="DarkSlateGray"
     VerticalAlignment="Stretch">
-
-    <Menu DockPanel.Dock="Top" Background="Purple" Padding="4" FontSize="14" FontFamily="Inter UI">
-      <MenuItem Header="_Game">
-        <MenuItem Header="_Host..."/>
-        <Separator/>
-        <MenuItem Header="_Join..."/>
-      </MenuItem>
-      <MenuItem Header="_Screen">
-        <MenuItem Header="_Ratio">
-          <MenuItem Header="_1920:1080"/>
-          <MenuItem Header="_800:640"/>
-          <MenuItem Header="_640:400"/>
-        </MenuItem>
-        <MenuItem Header="_Background..."/>
-      </MenuItem>
-      <MenuItem Header="_Audio">
-        <CheckBox BorderThickness="0"
-                  IsHitTestVisible="False"
-                  Command="{Binding ToggleCommand}">
-          Alert
-        </CheckBox>
-        <Separator/>
-        <CheckBox IsChecked="True" BorderThickness="0"
-                  IsHitTestVisible="False"
-                  Command="{Binding ToggleCommand}">
-          Music
-        </CheckBox>
-      </MenuItem>
-      <MenuItem Header="_Help">
-        <MenuItem Header="_About..." Command="{Binding AboutMenu}" CommandParameter="{Binding $parent[2]}"/>
-        <Separator/>
-        <MenuItem Header="_Version..."/>
-      </MenuItem>
-    </Menu>
     
     <StackPanel
       VerticalAlignment="Center">
