+++ Views/ToolbarMenuView.xaml.cs
@@ -1,0 +1,19 @@
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace Jeffistance.Views
+{
+    public class ToolbarMenuView : UserControl
+    {
+        public ToolbarMenuView()
+        {
+            InitializeComponent();
+        }
+
+        private void InitializeComponent()
+        {
+            AvaloniaXamlLoader.Load(this);
+        }
+    }
+}
